# Passthrough non-transformed strings (to some degree) in cssDecode