After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkCSSDecode
BenchmarkCSSDecode/#00
BenchmarkCSSDecode/#00-10         	463445419	         2.597 ns/op
BenchmarkCSSDecode/hello_world
BenchmarkCSSDecode/hello_world-10 	319812020	         3.715 ns/op
BenchmarkCSSDecode/test\a\b\f\n\r\t\v\?\'\"\\x00\12\123\1234\12345\123456\ff01\ff5e\_\\x00__string
BenchmarkCSSDecode/test\a\b\f\n\r\t\v\?\'\"\\x00\12\123\1234\12345\123456\ff01\ff5e\_\\x00__string-10         	 9009106	       133.5 ns/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/transformations
BenchmarkCSSDecode
BenchmarkCSSDecode/#00
BenchmarkCSSDecode/#00-10         	264846192	         4.522 ns/op
BenchmarkCSSDecode/hello_world
BenchmarkCSSDecode/hello_world-10 	52190552	        22.95 ns/op
BenchmarkCSSDecode/test\a\b\f\n\r\t\v\?\'\"\\x00\12\123\1234\12345\123456\ff01\ff5e\_\\x00__string
BenchmarkCSSDecode/test\a\b\f\n\r\t\v\?\'\"\\x00\12\123\1234\12345\123456\ff01\ff5e\_\\x00__string-10         	 8262039	       145.9 ns/op
PASS
```